script/CSRF SSRF XSS/requirements.txt
requests
beautifulsoup4
lxml
interactsh-client
selenium